+++ server/package.json
@@ -17,6 +17,7 @@
     "fastify-plugin": "^1.6.0",
     "fastify-redis": "^4.0.3",
     "fastify-sensible": "^2.1.1",
+    "fastify-websocket": "^2.0.5",
     "node-fetch": "^2.6.0",
     "socket.io": "2.3.0"
   },
